Move mapDesigns into separat ChangeNotifier
